546 commits from `next` -- both the dx and vulkan merges and the renderer-interface
reshuffle that came with them -- into a branch 292 commits of its own ahead. Only 7
files conflicted and 13 hunks, because this branch's work is almost entirely in its
own XNB-sweep area and misses the framework churn.

Six of the seven are the same 2026-09-06 dead-path and shared-ccache cleanup
wording resolved twice already today, and next carries the corrected history in
each: `build-probe/` is an in-repo directory rather than an "owner-designated
partition", the external /media/robertvokac/claude partition is gone rather than
current, and probe_selectors.sh gains both the explanation and the mandatory
CCACHE_BASEDIR. The seventh, CnbTextureCodecTests.cpp, is a one-line fixture-name
difference; the branch owner confirmed they had no stake in it.

Measured after the merge, from the repository root (these suites resolve fixtures
relative to the working directory -- running them from the build tree fails six
content-pipeline cases for that reason alone):

  CnaContentPipelineTests  481/481   -- exactly the pre-merge expectation
  CnaMathTests             854/854   -- 850 before; next adds four
  CnaContentTests          8 failed  -- 5 before
  CnaGraphicsTests         9 failed  -- 1 before

**No behavioural regression: nothing that passed before this merge fails after it.**
The eleven additional failures are all this branch's HEADLESS configuration meeting
tests written on next, which builds OPENGLES3 and never runs them there:

- Eight are in RenderTargetSemanticsTests.cpp, NpotTextureTests.cpp,
  PartialTextureTransfer and RenderTargetFormatAgreement -- files that did not
  exist on this branch before the merge. They read a render target back to the
  CPU, which HEADLESS refuses by name and always has.
- Two are the .cnj effect tests next changed in 0e2d80c45 (VULKAN-403) to require
  that a renderer which does not EXECUTE a shader source also reports the effect
  invalid. HEADLESS documentedly accepts any source and keeps drawing its own
  fixed path, so it answers valid; the test's implication does not hold there.
- One, XnbContentPipelineTest.SpriteFontRuntimeXnbAndTranscodedCnbHaveEquivalentSemantics,
  is already red on next itself and was imported with it.

So these are a negative-configuration gap in next's tests rather than damage from
merging: they need to ask the renderer's capability the way the LOD-bias and
compressed-format contracts already do. Recorded here rather than fixed, because
the fix belongs on next with all renderers in view, not inside this merge.

The frozen 7,726-reference corpus manifest under build/xna-sample-sweep/manifest/
is gitignored and exists only in this worktree; it was left untouched. The
differential sweep that would exercise this merge against the XNB pipeline is the
branch owner's to run.

+++ modules/renderers/directx11/include/CNA/Internal/Renderers/DirectX11/D3D11StateObjectCache.hpp
@@ -75,17 +75,16 @@ namespace CNA::Internal::Renderers::DirectX11
         std::map<Key, ComPtr<ID3D11DepthStencilState>> cache_;
     };
 
-    /// Caches ID3D11RasterizerState objects keyed by ApplyRasterizerState()'s 5 fields.
-    /// DepthBias is rounded to the nearest D3D11_RASTERIZER_DESC::DepthBias INT (see .cpp for the
-    /// unit-convention note); SlopeScaleDepthBias maps directly to the FLOAT
-    /// D3D11_RASTERIZER_DESC::SlopeScaledDepthBias field, no conversion needed.
+    /// Caches ID3D11RasterizerState objects keyed by ApplyRasterizerState()'s native fields.
+    /// The caller converts XNA's normalized DepthBias for the active DSV format before lookup;
+    /// SlopeScaleDepthBias maps directly to D3D11_RASTERIZER_DESC::SlopeScaledDepthBias.
     class D3D11RasterizerStateCache
     {
     public:
         ComPtr<ID3D11RasterizerState> GetOrCreate(ID3D11Device* device,
                                                    int cullMode, int fillMode,
                                                    bool scissorTestEnable,
-                                                   float depthBias, float slopeScaleDepthBias);
+                                                   int depthBias, float slopeScaleDepthBias);
 
         /// Number of distinct rasterizer states created so far (CNAEXT diagnostics).
         [[nodiscard]] std::size_t GetCacheSizeEXT() const { return cache_.size(); }
